+++ wmfdata/utils.py
@@ -166,4 +166,28 @@ def python_version():
     """
     return f"{sys.version_info.major}.{sys.version_info.minor}"
 
+def sql_tuple(i):
+    """
+    Given a Python iterable, returns a string representation that can be used in an SQL IN
+    clause.
 
+    For example:
+    > sql_tuple(["a", "b", "c"])
+    "('a', 'b', 'c')"
+    """
+    # It might seem useful to return "()" when an empty iterable is passed, but "IN ()"
+    # results in an SQL syntax error. Instead, we should send a clear signal to the caller
+    # that it needs to better handle the no-items case.
+    if len(i) == 0:
+        raise ValueError("Cannot produce an SQL tuple without any items.")
+
+    # Transform other iterables into lists, raising errors for non-iterables
+    if type(i) != list:
+        i = [x for x in i]
+
+    # Using Python's string representation functionality means we get escaping for free. Using the
+    # representation of a tuple almost works, but fails when there's just one element, because SQL
+    # doesn't accept the trailing comma that Python uses. Instead, we use representation of a
+    # list and replace the brackets with parentheses.
+    list_repr = repr(i)
+    return "(" + list_repr[1:-1] + ")"